# swagger.yaml
openapi: 3.0.0
info:
  title: JOBS API
  contact: {}
  version: '1.0'
servers:
  - url: https://jobshero1.herokuapp.com/api/v1
    variables: {}
paths:
  /jobs:
    post:
      tags:
        - Misc
      summary: Create jobs
      operationId: Createjobs
      parameters: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            encoding: {}
            schema:
              required:
                - position
                - company
              type: object
              properties:
                position:
                  type: string
                  example: manager
                company:
                  type: string
                  example: F1Soft
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
        - bearer: []
    get:
      tags:
        - Misc
      summary: getAlljobs
      operationId: getAlljobs
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
        - bearer: []
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...